APPCLI-43: Create hidden 'debug' command and shift 'configure info' to it
