+++ appcli/commands/configure_cli.py
@@ -160,25 +160,6 @@ class ConfigureCli:
                 # remove superfluous \n characters added by unified_diff
                 print(line.rstrip())
 
-        @configure.command(
-            hidden=True,
-            help="Prints detailed information about the current configuration.",
-        )
-        @click.pass_context
-        def info(ctx):
-            cli_context: CliContext = ctx.obj
-            print("=== CLI CONTEXT ===")
-            pprint(cli_context)
-            print("=== CONFIGURATION ===")
-            pprint(self.cli_configuration)
-            print("=== ORCHESTRATOR CONFIGURATION ===")
-            pprint(vars(self.cli_configuration.orchestrator))
-
-            app_config_file = cli_context.get_app_configuration_file()
-            variables_manager = VariablesManager(app_config_file)
-            print("=== VARIABLES ===")
-            pprint(variables_manager.get_all_variables())
-
         # Add the 'template' subcommand
         configure.add_command(ConfigureTemplateCli(self.cli_configuration).command)
